Update YA, fix get_categories

+++ market_api_app/base.py
@@ -28,7 +28,7 @@ class ApiBase:
             except requests.RequestException as e:
                 if attempt < self.max_retries - 1:
                     logger.debug(e.response.text)
-                    delay_seconds = float(e.response.headers.get('X-Ratelimit-Retry', self.delay_seconds))
+                    delay_seconds = int(e.response.headers.get('X-Ratelimit-Retry', self.delay_seconds))
                     logger.error(f'Неудачный запрос, ошибка: {e}. Повтор через {delay_seconds} секунд.')
                     time.sleep(delay_seconds)
                 else:
